chore: Add Pull Request Template (#23)
This pull request includes a small change to the
`.github/pull_request_template.md` file. The change adds a new template
for pull requests to ensure that descriptions are thorough and
consistent.

*
[`.github/pull_request_template.md`](diffhunk://#diff-b2496e80299b8c3150b1944450bd81c622e04e13d15c411d291db0927d75fd6bR1-R7):
Added a new template section for pull request descriptions to include a
summary, issue reference, motivation, context, and dependencies.

Fixes #1