ci: Add Markdown Link Checker (#49)
This pull request introduces a new workflow #19 eck Markdown links in
the repository and includes the necessary configuration for this check.

New workflow addition:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R129-R146):
Added a new job named `check-markdown-links` to perform link checking in
Markdown files using the `UmbrellaDocs/action-linkspector` action. This
job checks out the repository, runs the link checker, and reports any
issues as a GitHub pull request review.

Configuration for link checking:

*
[`.github/other-configurations/.linkspector.yml`](diffhunk://#diff-f9691f23ea6c4c34ecbaa23a98721e0aa3c7a94f2fcdd38f7b99fe863c5f6305R1-R4):
Added configuration for the link checker, specifying the directories to
check and the HTTP status codes considered as alive.

Fixes #19