ci: Add Linkinator Link Testing (#113)
# Pull Request

## Description

This pull request includes an addition to the GitHub Actions workflow
configuration in the `.github/workflows/deploy-to-github-pages.yml`
file. The most important change is the addition of a new job to run link
tests after deployment.

### Workflow Enhancements:

* Added a new job `link-tests` to run link tests using the
`JustinBeckwith/linkinator-action@v1.11.0` action. This job checks the
deployed site for broken links and ensures all links are functional.

Fixes #84