ci: Add run-scanner Job (#184)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to run a
scanner. The workflow is designed to handle debug mode, set up
dependencies, run the scanner, and upload the results. Below are the
most important changes:

New GitHub Actions workflow:

*
[`.github/workflows/run-scanner.yml`](diffhunk://#diff-b23f2b70f3c0ca075093fcdf8e636a7a731d588ac0f3d562b387140356d8642dR1-R44):
Added a new workflow named "Run Scanner" that includes a
`workflow_dispatch` trigger with an optional `scanner_debug` input. The
workflow sets read permissions for contents and includes a job to run
the scanner on `ubuntu-latest`. The job consists of steps to checkout
the repository, set up Python dependencies, run the scanner, rename the
results file, and upload the results as an artifact.

Fixes #183