build: Add Repositories Section (#186)
# Pull Request

## Description

This pull request includes changes to enhance the markdown output in the
`scanner/utils/write_markdown.py` and its associated tests. The changes
focus on improving the representation of technologies and frameworks by
adding badges and updating the structure of the markdown file.

Enhancements to markdown output:

*
[`scanner/utils/write_markdown.py`](diffhunk://#diff-345a5890da136100ba010968680627ae688839d4d186f1bf3191c1df15ad069fR22-R34):
Added a new section for repositories with badges for technologies and
frameworks.

Updates to tests:

*
[`scanner/utils/tests/test_write_markdown.py`](diffhunk://#diff-b69dfeeddcdeb2d19f21e2ede9823f430bd7fce0d49f1971dd5ea9401385e071L28-R35):
Updated the `test_write_output_file` and
`test_write_output_file__github_summary` tests to reflect the new
structure with badges for technologies and frameworks.
[[1]](diffhunk://#diff-b69dfeeddcdeb2d19f21e2ede9823f430bd7fce0d49f1971dd5ea9401385e071L28-R35)
[[2]](diffhunk://#diff-b69dfeeddcdeb2d19f21e2ede9823f430bd7fce0d49f1971dd5ea9401385e071R48-R64)
[[3]](diffhunk://#diff-b69dfeeddcdeb2d19f21e2ede9823f430bd7fce0d49f1971dd5ea9401385e071L68-R91)
[[4]](diffhunk://#diff-b69dfeeddcdeb2d19f21e2ede9823f430bd7fce0d49f1971dd5ea9401385e071L88-R120)

Fixes #179