build: Replace underscore with space in table headers (#192)
# Pull Request

## Description


This pull request includes updates to the GitHub Actions workflow for
code checks and an improvement to the markdown writing utility. The most
important changes are:

Updates to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L64-R64):
Updated the `github/codeql-action` to version `v3.28.1` for the
`upload-sarif`, `init`, and `analyze` steps.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L64-R64)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L112-R117)
[[3]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L186-R186)

Improvements to markdown writing utility:

*
[`scanner/utils/write_markdown.py`](diffhunk://#diff-345a5890da136100ba010968680627ae688839d4d186f1bf3191c1df15ad069fL77-R77):
Modified the `add_table` method to replace underscores with spaces in
table headers before capitalizing them.

Fixes #191