refactor: Refactor Pyproject (#195)
# Pull Request

## Description

This pull request includes several changes to the `Justfile` and
`pyproject.toml` files to improve dependency management and
configuration settings. The most important changes include updating the
installation command, modifying the project dependencies, and adjusting
configuration settings for the `ruff` linter and `vulture`.

Dependency Management:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL7-R7):
Updated the `install` command to include the `-E dev` option, ensuring
that development dependencies are also installed.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L3-L19):
Changed the dependency specification to use the `dynamic` field for the
version and added a `dependencies` section for the main dependencies and
an `optional-dependencies` section for development dependencies.

Configuration Settings:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L60-R72):
Adjusted the `dummy-variable-rgx` in the `ruff` linter configuration to
use a more precise regular expression.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L60-R72):
Modified the `convention` setting in the `pydocstyle` section to remove
the inline comment.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L60-R72):
Reformatted the `ignore_names` list in the `vulture` configuration for
better readability.