refactor: Python GitHub Actions (#196)
# Pull Request

## Description

This pull request includes several updates to the GitHub Actions
workflows and dependencies configuration. The key changes focus on
setting up dependencies and updating the development dependencies.

### Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L181-R184):
Replaced the `uv` setup step with a custom `setup-dependencies` action
and updated the `zizmor` command to use `poetry`.
*
[`.github/workflows/deploy-to-github-pages.yml`](diffhunk://#diff-ff49704c25d08d294e707839cc3d2831fbf4a6eada505c46b721163df00981eaR33-R34):
Added the `root-only` parameter to the `setup-dependencies` action.
*
[`.github/workflows/run-scanner.yml`](diffhunk://#diff-b23f2b70f3c0ca075093fcdf8e636a7a731d588ac0f3d562b387140356d8642dR32-R33):
Added the `root-only` parameter to the `setup-dependencies` action.

### Dependency updates:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L8-R8):
Added `zizmor==1.1.0` to the development dependencies.