test: Improve Markdown Tests to Check Contents (#203)
# Pull Request

## Description


This pull request includes updates to the GitHub Actions workflows and
enhancements to the markdown test for the tech report. The most
important changes involve renaming steps and artifacts in the workflows,
removing redundant tests, and adding new assertions in the test file.

Updates to GitHub Actions workflows:

*
[`.github/workflows/deploy-to-github-pages.yml`](diffhunk://#diff-ff49704c25d08d294e707839cc3d2831fbf4a6eada505c46b721163df00981eaL26-R26):
Renamed the "Checkout Repository" step to "Checkout" and the "Rename
Scanner Results" step to "mv tech_report.md index.md". Removed the
"Markdown Tests" job and updated dependencies.
[[1]](diffhunk://#diff-ff49704c25d08d294e707839cc3d2831fbf4a6eada505c46b721163df00981eaL26-R26)
[[2]](diffhunk://#diff-ff49704c25d08d294e707839cc3d2831fbf4a6eada505c46b721163df00981eaR38-R55)
[[3]](diffhunk://#diff-ff49704c25d08d294e707839cc3d2831fbf4a6eada505c46b721163df00981eaL90-L91)
[[4]](diffhunk://#diff-ff49704c25d08d294e707839cc3d2831fbf4a6eada505c46b721163df00981eaL121-R96)
*
[`.github/workflows/run-scanner.yml`](diffhunk://#diff-b23f2b70f3c0ca075093fcdf8e636a7a731d588ac0f3d562b387140356d8642dL25-R25):
Renamed the "Checkout Repository" step to "Checkout" and added a step to
rename the scanner results file. Removed the "Markdown Tests" job.
[[1]](diffhunk://#diff-b23f2b70f3c0ca075093fcdf8e636a7a731d588ac0f3d562b387140356d8642dL25-R25)
[[2]](diffhunk://#diff-b23f2b70f3c0ca075093fcdf8e636a7a731d588ac0f3d562b387140356d8642dR38-L67)

Enhancements to markdown test:

*
[`tests/markdown/test_tech_report.py`](diffhunk://#diff-be54cacaed9f5305cbd770d7eeba1b4962b3fa41ed46e298fb2e640e6bfbe441R3):
Added new headers and row count assertions for the summary and
repositories tables in the tech report.
[[1]](diffhunk://#diff-be54cacaed9f5305cbd770d7eeba1b4962b3fa41ed46e298fb2e640e6bfbe441R3)
[[2]](diffhunk://#diff-be54cacaed9f5305cbd770d7eeba1b4962b3fa41ed46e298fb2e640e6bfbe441R15-R46)

Fixes #202