ci: Update Dependency Review (#222)
# Pull Request

## Description

This pull request includes changes to the `jobs` section in the
`.github/workflows/pull-request-tasks.yml` file to enhance the
dependency review process. The most important changes include adding
permissions for pull requests, modifying the checkout step, and
configuring the dependency review action.

Enhancements to dependency review process:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR50-R61):
Added `pull-requests: write` permissions to the `dependency-review` job.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR50-R61):
Modified the checkout step to include `fetch-depth: 0` and
`persist-credentials: false`.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR50-R61):
Configured the dependency review action to comment a summary in the pull
request on failure.