ci: Validate Lefthook Configuration (#261)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow for
validating Lefthook configurations. The most significant change is the
inclusion of the `lefthook-validate` job in the
`.github/workflows/code-checks.yml` file.

### Workflow Enhancements:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R201-R216):
Added a new `lefthook-validate` job to the workflow. This job checks out
the repository, installs the latest version of `uv`, and runs `lefthook
validate` to ensure Lefthook configurations are valid.