ci: Add Pinact Verify Job (#265)
# Pull Request

## Description

This pull request introduces the integration of the Pinact tool into the
project for dependency verification and updates several configurations
to support its usage. It also includes minor updates to existing tools
and workflows.

### Pinact Integration:

* Added a new configuration file
`.github/other-configurations/pinact.yml` for Pinact, specifying the
schema and defining ignored actions.
* Updated `.github/workflows/code-checks.yml` to include a new
`pinact-verify` job for verifying dependencies using Pinact.
* Added Pinact-related commands (`pinact-run`, `pinact-check`,
`pinact-update`) to `Justfile` for running, verifying, and updating
dependencies.
* Updated `lefthook.yml` to include a `Pinact Checks` pre-commit hook,
which runs the `pinact-check` command.

### Tool and Workflow Updates:

* Bumped the minimum version of Lefthook in `lefthook.yml` from `1.11.3`
to `1.11.11`.
* Updated the `required-version` of `uv` in `pyproject.toml` from
`0.6.14` to `0.6.16`.