ci: Remove Pinact Verify Job (#269)
# Pull Request

## Description

This pull request includes updates to dependencies and workflow
configurations. The most important changes involve removing a redundant
workflow job in the GitHub Actions configuration and updating dependency
versions in `pyproject.toml` to ensure compatibility and access to the
latest features.

### Workflow configuration updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L216-L229):
Removed the `pinact-verify` job, which included steps for repository
checkout and installing a specific version of `uv`. This job was deemed
unnecessary.

### Dependency updates:
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L11-R11):
Updated the `ruff` dependency from version `0.11.6` to `0.11.7` for bug
fixes or improvements.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L22-R22):
Updated the `uv` required version from `0.6.16` to `0.6.17` to align
with the latest release.