deps(github-actions): Bump astral-sh/setup-uv to 6.0.1 (#272)
# Pull Request

## Description

This pull request updates the `setup-uv` action to a newer version
across multiple files to ensure compatibility with the latest features
and improvements.

### Updates to `setup-uv` Action:

*
[`.github/actions/setup-dependencies/action.yml`](diffhunk://#diff-41744aaf34820b46fe17d53d97ac7b6e435996a46c06a1781154ead9701846daL14-R14):
Updated the `setup-uv` action from version `v5.2.2` to `v6.0.1` to align
with the latest release. Removed the `pyproject-file` and `enable-cache`
configuration, likely due to changes in the action's interface.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L211-R211):
Updated the `setup-uv` action from version `v5.2.2` to `v6.0.1` in the
`Install the latest version of uv` step. Removed the `version`
parameter, reflecting updates in the action's configuration.