chore: Fix Zizmor SARIF Command (#284)
# Pull Request

## Description

This pull request updates the workflow and build tools to streamline
dependency management and improve the execution of code checks. The most
important changes include replacing the Python dependency setup with new
tools (`uv` and `Just`) and updating the commands for running and
uploading SARIF reports.

### Workflow updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L187-R196):
Replaced the Python dependency setup action with `astral-sh/setup-uv`
and `extractions/setup-just` to install `uv` and set up `Just`. Updated
the `Run zizmor` step to use the `just zizmor-check-sarif` command and
upgraded the SARIF upload action to `v3.28.17`.

### Build tool updates:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL144-R148):
Added a new `zizmor-check-sarif` command using `uvx` to generate SARIF
output for the `zizmor` checks. Updated the existing `zizmor-check`
command to use `uvx` instead of the previous `zizmor` command.