chore: Remove Zizmor Locked Version (#285)
# Pull Request

## Description

This pull request updates dependencies in the `pyproject.toml` file to
ensure compatibility and take advantage of the latest features and
fixes. The most important changes involve version bumps for several
libraries and tools.

### Dependency Updates:
* Updated `structlog` from `25.2.0` to `25.3.0` to include the latest
improvements and bug fixes.
* Updated `pytest-cov` from `6.1.0` to `6.1.1` and `ruff` from `0.11.7`
to `0.11.9` in the `dev` optional dependencies for development tooling
enhancements.
* Updated `uv` required version from `0.6.17` to `0.7.3` to ensure
compatibility with newer versions.