Swap SuperLinter to Full Version (#290)
# Pull Request

## Description

This pull request includes a minor update to the
`.github/workflows/code-checks.yml` file. The change updates the GitHub
Action for linting the codebase to use the standard
`super-linter/super-linter` repository instead of the
`super-linter/super-linter/slim` variant.

* Workflow update:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L27-R28):
Updated the `Lint Code Base` step to use `super-linter/super-linter`
instead of `super-linter/super-linter/slim` for consistency and to align
with the standard repository.