ci: Bump SuperLinter to 7.4.0 (#293)
# Pull Request

## Description

This pull request updates the linter configuration in the GitHub Actions
workflow to use a newer version of the `super-linter` and removes an
unused environment variable.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L28-L35):
Updated `super-linter` to version `7.4.0` by changing the commit hash in
the `uses` field.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L28-L35):
Removed the `EDITORCONFIG_FILE_NAME` environment variable, as it is no
longer required.