test: Add More UI Tests (#305)
# Pull Request

## Description

This pull request adds several new test cases to the
`tests/ui/test_website.py` file to improve test coverage for the web
application's UI. The tests focus on verifying the presence and
structure of headers and tables on the main project page.

### Added test cases for UI verification:

* **Header tests**: Added `test_headers` to verify that the page
contains the correct `h1` headers and that the "source_scan" text links
to the correct URL.
* **Table count test**: Added `test_for_2_tables` to ensure that the
page contains exactly two tables.
* **Summary table tests**: Added `test_summary_table` to validate the
structure and content of the summary table, including the number of
rows, headers, and their text.
* **Repositories table tests**: Added `test_repositories_table` to
validate the structure and content of the repositories table, including
the number of rows, headers, and their text.

Fixes #240