chore: Migrate diagrams code to drawio png (#327)
# Pull Request

## Description

This pull request removes the `c4.py` diagram script, which previously
generated a C4 model diagram for the project. The most important change
is the deletion of the diagram generation code and its associated
elements.

### Diagram removal:

*
[`diagrams/c4.py`](diffhunk://#diff-a6b4e89a2a3973dd3e2818b3f97dd505f45a853fcc7123a82a850534dcca4ab1L1-L23):
Removed the entire script, including imports from the `diagrams`
library, definitions for `Person`, `Container`, `Relationship`, and
`SystemBoundary`, and the logic for creating a C4 model diagram.

Fixes #325