chore: Remove Zizmor Command (#332)
# Pull Request

## Description

This pull request removes an unused target from the `Justfile`. The
removed target, `zizmor-check-sarif`, was used to run `zizmor` with
SARIF output, but it is no longer needed.

Code cleanup:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL164-L167):
Removed the `zizmor-check-sarif` target, which ran `zizmor` with SARIF
output and saved the results to a file, as it is now redundant.