chore: Set Zizmor persona to auditor (#337)
# Pull Request

## Description

This pull request updates several workflow files to use newer versions
of reusable workflows and actions, adjusts a configuration in the
`Justfile`, and updates the minimum version for `lefthook`. The changes
primarily focus on maintaining compatibility and ensuring the use of the
latest versions for improved functionality and security.

### Workflow Updates
*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reusable workflow to version `v2025.06.16.01`.
* `.github/workflows/code-checks.yml`: 
  - Updated the reusable workflow to version `v2025.06.16.01`.
- Updated the `github/codeql-action/upload-sarif` action to version
`v3.29.0`.
- Updated the `codeql-analysis` reusable workflow to version
`v2025.06.16.01`.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL24-R24):
Updated the reusable workflow to version `v2025.06.16.01`.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L18-R18):
Updated the reusable workflow to version `v2025.06.16.01`.

### Configuration Changes
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL162-R162):
Changed the `zizmor-check` persona from `pedantic` to `auditor` for
validation purposes.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the minimum version of `lefthook` from `1.11.11` to `1.11.14`.