ci: Temporarily Disable Type Checks (#340)
# Pull Request

## Description

This pull request includes changes to improve code reliability,
streamline workflows, and enhance type checking. The most important
changes involve error handling for environment variables, commenting out
a workflow job, and adding type annotations for better code clarity.

### Code reliability improvements:
*
[`tests/ui/variables.py`](diffhunk://#diff-63926a38ac6210e0eb1284da014076012c396be52f9a92ae3090b88fc74781b4R4-R6):
Added a check to ensure the `PROJECT_URL` environment variable is set,
raising a `ValueError` with a descriptive message if it is not.

### Workflow streamlining:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L104-R118):
Commented out the `run-python-type-checks` job to temporarily disable
Python type checking in the CI pipeline.

### Type checking enhancements:
*
[`scanner/utils/github_interactions.py`](diffhunk://#diff-40dae3e48e37722006554d3f1db64df58148a03b169aeda6aad8d41bc63d8859R12-R13):
Added type annotations for the `Repository` class from the `github`
module to improve code clarity and maintainability.