chore: Fix UI Test (#358)
# Pull Request

## Description

This pull request includes updates to dependencies in `pyproject.toml`
and a test adjustment in `tests/ui/test_website.py` to reflect a change
in the number of rows in a table. Below is a summary of the most
important changes:

### Dependency Updates:
* Updated `ruff` from version `0.12.1` to `0.12.3` in `pyproject.toml`.
* Updated `ty` from version `0.0.1a12` to `0.0.1a14` in
`pyproject.toml`.
* Updated `required-version` of `uv` from `0.7.16` to `0.7.20` in
`pyproject.toml`.

### Test Adjustments:
* Updated the assertion in `test_repositories_table` to expect 34 rows
instead of 33 in `tests/ui/test_website.py`.