chore: Update UV Version (#368)
# Pull Request

## Description

This pull request updates the required version for the `uv` tool in the
`pyproject.toml` file to ensure compatibility with newer features and
bug fixes.

Version update:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L22-R22):
Changed the `required-version` of the `uv` tool from `"0.7.20"` to
`"~=0.8.0"`. This allows the project to use any version of `uv` starting
from `0.8.0` but below `0.9.0`.