chore: Add actionlint Git Hook (#385)
# Pull Request

## Description

This pull request introduces a new `actionlint` check to the development
workflow, ensuring that GitHub Actions workflows are validated for
correctness. The changes include adding the `actionlint` command to the
`Justfile` and integrating the check into the `lefthook` pre-commit
hooks.

### Additions to development workflow:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR171-R178):
Added a new `actionlint-check` target that runs the `actionlint` command
to validate GitHub Actions workflows.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR25-R26):
Integrated the `actionlint-check` target into the `pre-commit` hooks
under the name "Actionlint Checks."